requirements.txt
setuptools==75.5.0
uvicorn==0.32.0
starlette==0.41.2
python-dotenv==1.0.1
openai==1.54.4
asqlite==2.0.0
requests~=2.32.3
python-i18n==0.3.9